requirements.txt
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
chromadb>=0.4.22
pypdf>=4.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.2

